First Test - Add service and test for removing a user from a team
Implemented the `eliminarUsuarioDeEquipo` method in `EquipoService` to remove a user from a team. This method checks if both the team and the user exist, throwing `EquipoServiceException` if any of them are not found. It updates both sides of the bidirectional relationship.

Also added a new test `eliminarUsuarioDeEquipoTest` in `EquipoServiceTest` to validate the correct behavior:
- Registers a user and creates a team.
- Adds the user to the team.
- Removes the user from the team using the new method.
- Asserts that the user is no longer part of the team.

These changes improve the system's ability to manage team memberships dynamically.