Third Test - Prevent Duplicate User Additions to Team
Refactored the `añadirUsuarioAEquipo` method in `EquipoService` to prevent duplicate user entries in the same team. Now, if a user already belongs to the team, an `EquipoServiceException` is thrown. This improves data consistency and prevents redundant associations.

Also added the corresponding test `añadirUsuarioDuplicadoAEquipoTest` in `EquipoServiceTest` to verify that trying to add an already existing user to a team throws the expected exception.